[TECH] Fix seeds dependency on migration data
Context
-------

The `emptyAllTables` method clears all data in the database, including
data generated by the migration scripts.

Some seeds depend on the data generated by the migration scripts, like
the `PIX_MASTER` role, the `ADMIN` organization role, and for the
campaigns, a target profile.

Problem
-------

Applying the `emptyAllTables` method removes everything including the
migration data, so applying the seeds again with Postgres fails because
of missing foreign keys.

Fix
---

This fix includes the creation of all the needed data in the seeds.