[TECH] Removes the random port when starting the API with `test` env.
Context
-------

When starting the API whith `NODE_ENV=test`, the behaviour was to ignore
any `PORT` configuration and to start the server on a random available
port.

Problem
-------

To test with Cypress, we whish to start the API with `NODE_ENV=test` and
call it on the configured port, like `3000`.

Fix
---

This fix removes the random `port` behaviour: the API now starts on the
port specified by the configuration, even when `NODE_ENV=test`.